+++ Server.ts
@@ -9,6 +9,7 @@ import * as schedule from "node-schedule";
 import * as cookieParser from "cookie-parser";
 
 import { CookieParserConfig } from "./api/config/CookieParserConfig";
+import { ClearTokenJob } from "./jobs/ClearTokens/ClearTokenJob";
 
 let AuthRoutes = require("./api/Routes/AuthRoutes");
 let UserRoutes = require("./api/Routes/UserRoutes");
@@ -144,14 +145,9 @@ export class Server
         // });
 
         // example rule
-        let rule = new schedule.RecurrenceRule();
-        rule.dayOfWeek = [ new schedule.Range(0, 6) ];
-        rule.hour = [ new schedule.Range(0, 23) ];
-        rule.minute = [ new schedule.Range(0, 59) ];
-
-        let job = schedule.scheduleJob(rule, function() {
-            console.log('My name is Jobert Jith');
-        });
+        
+        ClearTokenJob.register(this.db);
+        
 
     }
