+++ npm-debug.log
@@ -10,10 +10,10 @@
 9 verbose lifecycle server@0.0.0~compile: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/usr/src/app/node_modules/.bin:/usr/local/lib/node_modules/npm/bin/node-gyp-bin:/usr/src/app/node_modules/.bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
 10 verbose lifecycle server@0.0.0~compile: CWD: /usr/src/app
 11 silly lifecycle server@0.0.0~compile: Args: [ '-c', 'tsc && node ./bin/www' ]
-12 silly lifecycle server@0.0.0~compile: Returned: code: 1  signal: null
+12 silly lifecycle server@0.0.0~compile: Returned: code: 2  signal: null
 13 info lifecycle server@0.0.0~compile: Failed to exec compile script
 14 verbose stack Error: server@0.0.0 compile: `tsc && node ./bin/www`
-14 verbose stack Exit status 1
+14 verbose stack Exit status 2
 14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:279:16)
 14 verbose stack     at emitTwo (events.js:106:13)
 14 verbose stack     at EventEmitter.emit (events.js:192:7)
@@ -30,7 +30,7 @@
 20 error npm  v4.1.2
 21 error code ELIFECYCLE
 22 error server@0.0.0 compile: `tsc && node ./bin/www`
-22 error Exit status 1
+22 error Exit status 2
 23 error Failed at the server@0.0.0 compile script 'tsc && node ./bin/www'.
 23 error Make sure you have the latest version of node.js and npm installed.
 23 error If you do, this is most likely a problem with the server package,
